+++ src/inmanta/protocol/methods.py
@@ -251,6 +251,12 @@ def decomission_environment(id: uuid.UUID, metadata: dict = None):
     """
     Decommision an environment. This is done by uploading an empty model to the server and let purge_on_delete handle
     removal.
+
+    :param id: The uuid of the environment.
+    :param metadata: Optional metadata associated with the decommissioning
+
+    :raises NotFound: The given environment doesn't exist.
+    :raises Forbidden: The given environment is protected.
     """