+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Fix version cli argument conflict (#2358)
 - Don't remove resource independent parameters on version deletion (#2370)
 - Enhance installation documentation (#2241, #2356, #2357)
+- Ensure that a protected environment can't be decommissioned (#2376)
 
 # Release 2020.4 (2020-09-08)